- Restructured fragments that graph things under umbrella of GraphViewFragment - Repositioned XML to look better - Added Anterior/Posterior Fragment - Added Title for each view
